requirements.txt
flask==3.0.0
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.5.0
# Testing dependencies
coverage==7.3.2